Fixed screen output
vasprintf() breaks the output for me (on Brahma)

Clearing the screen makes sense, because some ncchinfo.bin files
completely fill the screen with Debug messages, leading to bad output
otherwise.

current_y should not be modified externally, but you may want to put it
back in if you intended to use that.