Enable working directory handling
This is designed to prevent confusion (with old tutorials) and still
allow full usage of a working directory.

If /Decrypt9 exists, it is used as standard output directory and files
are first searched in it, then in the root dir. If /Decrypt9 does not
exist, everything works as before.

This commit also changes the decrypt directory for NCCH/NCSD to
'/D9Game', which is more consistent with the name of the feature in the
menu. '/Decrypt9', if it exists, now also works as a second alternative
to '/D9Game'.

This commit also adds some coding style fixes to fs.c