Added ability to encrypt/decrypt 3DS SD files
This is an alternative to SD padgen. It works by first copying the file
structure (or part of it) from 'SD://Nintendo 3DS/ID0/ID1/' to the game
directory ('D9Game') or the work directory ('Decrypt9'). It doesn't process the 'backup' and 'DSiWare' folders, because these use a different kind of encryption.

Because there is no way for us to safely determine wether files are
encrypted or not (apart from statistical analysis), this also works as
an encryptor.

Also, the DecryptSdToSd() function is now called CryptSdToSd, because it
also does encryption.