Initial plug-in of new allocator, re-write of linker script
Unresolved question: apparently newlib also provides a malloc function, which is currently hooked up, but how it works seems pretty magical to me.

We should figure out what its heap space is and how it might conflict with this new allocator, as well as whether it would be easier/more worth it to use just the newlib allocator.

Currently, malloc is used internally by a number of libc functions (to my surprise).